Refect: NotionColor로 색 입히는 부분 모듈화해서 빼기 (TODO 처리)
